Move upsert/delete e-service template functions to `eservice-template-rmw-sql`
