+++ packages/eservice-template-readmodel-writer-sql/package.json
@@ -34,6 +34,7 @@
     "@protobuf-ts/runtime": "2.9.4",
     "connection-string": "4.4.0",
     "dotenv-flow": "4.1.0",
+    "drizzle-orm": "0.39.3",
     "kafka-iam-auth": "workspace:*",
     "kafkajs": "2.2.4",
     "pagopa-interop-commons": "workspace:*",
