+++ packages/eservice-template-readmodel-writer-sql/test/eserviceTemplateReadmodelWriter.integration.test.ts
@@ -35,13 +35,16 @@ import {
   tenantKind,
 } from "pagopa-interop-models";
 import { handleMessageV2 } from "../src/consumerServiceV2.js";
-import { eserviceTemplateReadModelService } from "./utils.js";
+import {
+  eserviceTemplateReadModelService,
+  eserviceTemplateWriterService,
+} from "./utils.js";
 
 describe("database test", async () => {
   describe("Events V2", async () => {
     const mockEServiceTemplate = getMockEServiceTemplate();
     it("EServiceTemplateDeleted", async () => {
-      await eserviceTemplateReadModelService.upsertEServiceTemplate(
+      await eserviceTemplateWriterService.upsertEServiceTemplate(
         mockEServiceTemplate,
         1
       );
@@ -58,7 +61,7 @@ describe("database test", async () => {
         data: payload,
         log_date: new Date(),
       };
-      await handleMessageV2(message, eserviceTemplateReadModelService);
+      await handleMessageV2(message, eserviceTemplateWriterService);
 
       const retrievedEservice =
         await eserviceTemplateReadModelService.getEServiceTemplateById(
@@ -81,7 +84,7 @@ describe("database test", async () => {
         data: payload,
         log_date: new Date(),
       };
-      await handleMessageV2(message, eserviceTemplateReadModelService);
+      await handleMessageV2(message, eserviceTemplateWriterService);
 
       const retrievedEservice =
         await eserviceTemplateReadModelService.getEServiceTemplateById(
@@ -95,7 +98,7 @@ describe("database test", async () => {
     });
 
     it("EServiceTemplateDraftUpdated", async () => {
-      await eserviceTemplateReadModelService.upsertEServiceTemplate(
+      await eserviceTemplateWriterService.upsertEServiceTemplate(
         mockEServiceTemplate,
         1
       );
@@ -118,7 +121,7 @@ describe("database test", async () => {
         data: payload,
         log_date: new Date(),
       };
-      await handleMessageV2(message, eserviceTemplateReadModelService);
+      await handleMessageV2(message, eserviceTemplateWriterService);
 
       const retrievedEservice =
         await eserviceTemplateReadModelService.getEServiceTemplateById(
@@ -140,7 +143,7 @@ describe("database test", async () => {
         ...mockEServiceTemplate,
         versions: [],
       };
-      await eserviceTemplateReadModelService.upsertEServiceTemplate(
+      await eserviceTemplateWriterService.upsertEServiceTemplate(
         eserviceTemplate,
         1
       );
@@ -162,7 +165,7 @@ describe("database test", async () => {
         data: payload,
         log_date: new Date(),
       };
-      await handleMessageV2(message, eserviceTemplateReadModelService);
+      await handleMessageV2(message, eserviceTemplateWriterService);
 
       const retrievedEservice =
         await eserviceTemplateReadModelService.getEServiceTemplateById(
@@ -182,7 +185,7 @@ describe("database test", async () => {
         ...mockEServiceTemplate,
         versions: [draftVersion],
       };
-      await eserviceTemplateReadModelService.upsertEServiceTemplate(
+      await eserviceTemplateWriterService.upsertEServiceTemplate(
         eserviceTemplate,
         1
       );
@@ -204,7 +207,7 @@ describe("database test", async () => {
         data: payload,
         log_date: new Date(),
       };
-      await handleMessageV2(message, eserviceTemplateReadModelService);
+      await handleMessageV2(message, eserviceTemplateWriterService);
 
       const retrievedEservice =
         await eserviceTemplateReadModelService.getEServiceTemplateById(
@@ -227,7 +230,7 @@ describe("database test", async () => {
         ...mockEServiceTemplate,
         versions: [draftVersion],
       };
-      await eserviceTemplateReadModelService.upsertEServiceTemplate(
+      await eserviceTemplateWriterService.upsertEServiceTemplate(
         eserviceTemplate,
         1
       );
@@ -253,7 +256,7 @@ describe("database test", async () => {
         data: payload,
         log_date: new Date(),
       };
-      await handleMessageV2(message, eserviceTemplateReadModelService);
+      await handleMessageV2(message, eserviceTemplateWriterService);
 
       const retrievedEservice =
         await eserviceTemplateReadModelService.getEServiceTemplateById(
@@ -275,7 +278,7 @@ describe("database test", async () => {
         ...mockEServiceTemplate,
         versions: [publishedVersion],
       };
-      await eserviceTemplateReadModelService.upsertEServiceTemplate(
+      await eserviceTemplateWriterService.upsertEServiceTemplate(
         eserviceTemplate,
         1
       );
@@ -302,7 +305,7 @@ describe("database test", async () => {
         data: payload,
         log_date: new Date(),
       };
-      await handleMessageV2(message, eserviceTemplateReadModelService);
+      await handleMessageV2(message, eserviceTemplateWriterService);
 
       const retrievedEservice =
         await eserviceTemplateReadModelService.getEServiceTemplateById(
@@ -325,7 +328,7 @@ describe("database test", async () => {
         ...mockEServiceTemplate,
         versions: [suspendedEServiceTemplateVersion],
       };
-      await eserviceTemplateReadModelService.upsertEServiceTemplate(
+      await eserviceTemplateWriterService.upsertEServiceTemplate(
         eserviceTemplate,
         1
       );
@@ -353,7 +356,7 @@ describe("database test", async () => {
         data: payload,
         log_date: new Date(),
       };
-      await handleMessageV2(message, eserviceTemplateReadModelService);
+      await handleMessageV2(message, eserviceTemplateWriterService);
 
       const retrievedEservice =
         await eserviceTemplateReadModelService.getEServiceTemplateById(
@@ -374,7 +377,7 @@ describe("database test", async () => {
         ...mockEServiceTemplate,
         versions: [draftEServiceTemplateVersion],
       };
-      await eserviceTemplateReadModelService.upsertEServiceTemplate(
+      await eserviceTemplateWriterService.upsertEServiceTemplate(
         eserviceTemplate,
         1
       );
@@ -401,7 +404,7 @@ describe("database test", async () => {
         data: payload,
         log_date: new Date(),
       };
-      await handleMessageV2(message, eserviceTemplateReadModelService);
+      await handleMessageV2(message, eserviceTemplateWriterService);
 
       const retrievedEservice =
         await eserviceTemplateReadModelService.getEServiceTemplateById(
@@ -423,7 +426,7 @@ describe("database test", async () => {
         ...mockEServiceTemplate,
         versions: [publishedEServiceTemplateVersion],
       };
-      await eserviceTemplateReadModelService.upsertEServiceTemplate(
+      await eserviceTemplateWriterService.upsertEServiceTemplate(
         eserviceTemplate,
         1
       );
@@ -450,7 +453,7 @@ describe("database test", async () => {
         data: payload,
         log_date: new Date(),
       };
-      await handleMessageV2(message, eserviceTemplateReadModelService);
+      await handleMessageV2(message, eserviceTemplateWriterService);
 
       const retrievedEservice =
         await eserviceTemplateReadModelService.getEServiceTemplateById(
@@ -472,7 +475,7 @@ describe("database test", async () => {
         ...mockEServiceTemplate,
         versions: [draftEServiceTemplateVersion],
       };
-      await eserviceTemplateReadModelService.upsertEServiceTemplate(
+      await eserviceTemplateWriterService.upsertEServiceTemplate(
         eserviceTemplate,
         1
       );
@@ -496,7 +499,7 @@ describe("database test", async () => {
         data: payload,
         log_date: new Date(),
       };
-      await handleMessageV2(message, eserviceTemplateReadModelService);
+      await handleMessageV2(message, eserviceTemplateWriterService);
 
       const retrievedEservice =
         await eserviceTemplateReadModelService.getEServiceTemplateById(
@@ -518,7 +521,7 @@ describe("database test", async () => {
         ...mockEServiceTemplate,
         versions: [draftEServiceTemplateVersion],
       };
-      await eserviceTemplateReadModelService.upsertEServiceTemplate(
+      await eserviceTemplateWriterService.upsertEServiceTemplate(
         eserviceTemplate,
         1
       );
@@ -541,7 +544,7 @@ describe("database test", async () => {
         data: payload,
         log_date: new Date(),
       };
-      await handleMessageV2(message, eserviceTemplateReadModelService);
+      await handleMessageV2(message, eserviceTemplateWriterService);
 
       const retrievedEservice =
         await eserviceTemplateReadModelService.getEServiceTemplateById(
@@ -563,7 +566,7 @@ describe("database test", async () => {
         ...mockEServiceTemplate,
         versions: [draftEServiceTemplateVersion],
       };
-      await eserviceTemplateReadModelService.upsertEServiceTemplate(
+      await eserviceTemplateWriterService.upsertEServiceTemplate(
         eserviceTemplate,
         1
       );
@@ -592,7 +595,7 @@ describe("database test", async () => {
         data: payload,
         log_date: new Date(),
       };
-      await handleMessageV2(message, eserviceTemplateReadModelService);
+      await handleMessageV2(message, eserviceTemplateWriterService);
 
       const retrievedEservice =
         await eserviceTemplateReadModelService.getEServiceTemplateById(
@@ -614,7 +617,7 @@ describe("database test", async () => {
         ...mockEServiceTemplate,
         versions: [draftEServiceTemplateVersion],
       };
-      await eserviceTemplateReadModelService.upsertEServiceTemplate(
+      await eserviceTemplateWriterService.upsertEServiceTemplate(
         eserviceTemplate,
         1
       );
@@ -643,7 +646,7 @@ describe("database test", async () => {
         data: payload,
         log_date: new Date(),
       };
-      await handleMessageV2(message, eserviceTemplateReadModelService);
+      await handleMessageV2(message, eserviceTemplateWriterService);
 
       const retrievedEservice =
         await eserviceTemplateReadModelService.getEServiceTemplateById(
@@ -665,7 +668,7 @@ describe("database test", async () => {
         ...mockEServiceTemplate,
         versions: [draftEServiceTemplateVersion],
       };
-      await eserviceTemplateReadModelService.upsertEServiceTemplate(
+      await eserviceTemplateWriterService.upsertEServiceTemplate(
         eserviceTemplate,
         1
       );
@@ -694,7 +697,7 @@ describe("database test", async () => {
         data: payload,
         log_date: new Date(),
       };
-      await handleMessageV2(message, eserviceTemplateReadModelService);
+      await handleMessageV2(message, eserviceTemplateWriterService);
 
       const retrievedEservice =
         await eserviceTemplateReadModelService.getEServiceTemplateById(
@@ -716,7 +719,7 @@ describe("database test", async () => {
         ...mockEServiceTemplate,
         versions: [draftEServiceTemplateVersion],
       };
-      await eserviceTemplateReadModelService.upsertEServiceTemplate(
+      await eserviceTemplateWriterService.upsertEServiceTemplate(
         eserviceTemplate,
         1
       );
@@ -740,7 +743,7 @@ describe("database test", async () => {
         data: payload,
         log_date: new Date(),
       };
-      await handleMessageV2(message, eserviceTemplateReadModelService);
+      await handleMessageV2(message, eserviceTemplateWriterService);
 
       const retrievedEservice =
         await eserviceTemplateReadModelService.getEServiceTemplateById(
@@ -752,7 +755,7 @@ describe("database test", async () => {
     });
 
     it("EServiceTemplateRiskAnalysisAdded", async () => {
-      await eserviceTemplateReadModelService.upsertEServiceTemplate(
+      await eserviceTemplateWriterService.upsertEServiceTemplate(
         mockEServiceTemplate,
         1
       );
@@ -777,7 +780,7 @@ describe("database test", async () => {
         data: payload,
         log_date: new Date(),
       };
-      await handleMessageV2(message, eserviceTemplateReadModelService);
+      await handleMessageV2(message, eserviceTemplateWriterService);
 
       const retrievedEservice =
         await eserviceTemplateReadModelService.getEServiceTemplateById(
@@ -798,7 +801,7 @@ describe("database test", async () => {
         ...mockEServiceTemplate,
         riskAnalysis: [mockRiskAnalysis],
       };
-      await eserviceTemplateReadModelService.upsertEServiceTemplate(
+      await eserviceTemplateWriterService.upsertEServiceTemplate(
         eserviceTemplate,
         1
       );
@@ -833,7 +836,7 @@ describe("database test", async () => {
         data: payload,
         log_date: new Date(),
       };
-      await handleMessageV2(message, eserviceTemplateReadModelService);
+      await handleMessageV2(message, eserviceTemplateWriterService);
 
       const retrievedEservice =
         await eserviceTemplateReadModelService.getEServiceTemplateById(
@@ -855,7 +858,7 @@ describe("database test", async () => {
         riskAnalysis: [riskAnalysis],
       };
 
-      await eserviceTemplateReadModelService.upsertEServiceTemplate(
+      await eserviceTemplateWriterService.upsertEServiceTemplate(
         eserviceTemplate,
         1
       );
@@ -877,7 +880,7 @@ describe("database test", async () => {
         data: payload,
         log_date: new Date(),
       };
-      await handleMessageV2(message, eserviceTemplateReadModelService);
+      await handleMessageV2(message, eserviceTemplateWriterService);
 
       const retrievedEservice =
         await eserviceTemplateReadModelService.getEServiceTemplateById(
@@ -902,7 +905,7 @@ describe("database test", async () => {
         name: "previousName",
         versions: [publishedEServiceTemplateVersion],
       };
-      await eserviceTemplateReadModelService.upsertEServiceTemplate(
+      await eserviceTemplateWriterService.upsertEServiceTemplate(
         eserviceTemplate,
         1
       );
@@ -922,7 +925,7 @@ describe("database test", async () => {
         data: payload,
         log_date: new Date(),
       };
-      await handleMessageV2(message, eserviceTemplateReadModelService);
+      await handleMessageV2(message, eserviceTemplateWriterService);
       const retrievedEservice =
         await eserviceTemplateReadModelService.getEServiceTemplateById(
           mockEServiceTemplate.id
