+++ packages/readmodel/src/testUtils.ts
@@ -1,5 +1,6 @@
 import {
   Agreement,
+  EServiceTemplate,
   TenantNotificationConfig,
   UserNotificationConfig,
 } from "pagopa-interop-models";
@@ -10,6 +11,13 @@ import {
   agreementInReadmodelAgreement,
   agreementStampInReadmodelAgreement,
   DrizzleReturnType,
+  eserviceTemplateInReadmodelEserviceTemplate,
+  eserviceTemplateRiskAnalysisAnswerInReadmodelEserviceTemplate,
+  eserviceTemplateRiskAnalysisInReadmodelEserviceTemplate,
+  eserviceTemplateVersionAttributeInReadmodelEserviceTemplate,
+  eserviceTemplateVersionDocumentInReadmodelEserviceTemplate,
+  eserviceTemplateVersionInReadmodelEserviceTemplate,
+  eserviceTemplateVersionInterfaceInReadmodelEserviceTemplate,
   tenantNotificationConfigInReadmodelNotificationConfig,
   userNotificationConfigInReadmodelNotificationConfig,
 } from "pagopa-interop-readmodel-models";
@@ -20,6 +28,7 @@ import {
 } from "./notification-config/splitters.js";
 import { splitAgreementIntoObjectsSQL } from "./agreement/splitters.js";
 import { checkMetadataVersion } from "./utils.js";
+import { splitEServiceTemplateIntoObjectsSQL } from "./eservice-template/splitters.js";
 
 export const insertTenantNotificationConfig = async (
   readModelDB: DrizzleReturnType,
@@ -101,6 +110,81 @@ export const upsertAgreement = async (
       await tx
         .insert(agreementContractInReadmodelAgreement)
         .values(contractSQL);
+    }
+  });
+};
+
+export const upsertEServiceTemplate = async (
+  readModelDB: DrizzleReturnType,
+  eserviceTemplate: EServiceTemplate,
+  metadataVersion: number
+): Promise<void> => {
+  await readModelDB.transaction(async (tx) => {
+    const shouldUpsert = await checkMetadataVersion(
+      tx,
+      eserviceTemplateInReadmodelEserviceTemplate,
+      metadataVersion,
+      eserviceTemplate.id
+    );
+
+    if (!shouldUpsert) {
+      return;
+    }
+
+    await tx
+      .delete(eserviceTemplateInReadmodelEserviceTemplate)
+      .where(
+        eq(eserviceTemplateInReadmodelEserviceTemplate.id, eserviceTemplate.id)
+      );
+
+    const {
+      eserviceTemplateSQL,
+      riskAnalysesSQL,
+      riskAnalysisAnswersSQL,
+      versionsSQL,
+      attributesSQL,
+      interfacesSQL,
+      documentsSQL,
+    } = splitEServiceTemplateIntoObjectsSQL(eserviceTemplate, metadataVersion);
+
+    await tx
+      .insert(eserviceTemplateInReadmodelEserviceTemplate)
+      .values(eserviceTemplateSQL);
+
+    for (const versionSQL of versionsSQL) {
+      await tx
+        .insert(eserviceTemplateVersionInReadmodelEserviceTemplate)
+        .values(versionSQL);
+    }
+
+    for (const interfaceSQL of interfacesSQL) {
+      await tx
+        .insert(eserviceTemplateVersionInterfaceInReadmodelEserviceTemplate)
+        .values(interfaceSQL);
+    }
+
+    for (const docSQL of documentsSQL) {
+      await tx
+        .insert(eserviceTemplateVersionDocumentInReadmodelEserviceTemplate)
+        .values(docSQL);
+    }
+
+    for (const attributeSQL of attributesSQL) {
+      await tx
+        .insert(eserviceTemplateVersionAttributeInReadmodelEserviceTemplate)
+        .values(attributeSQL);
+    }
+
+    for (const riskAnalysisSQL of riskAnalysesSQL) {
+      await tx
+        .insert(eserviceTemplateRiskAnalysisInReadmodelEserviceTemplate)
+        .values(riskAnalysisSQL);
+    }
+
+    for (const riskAnalysisAnswerSQL of riskAnalysisAnswersSQL) {
+      await tx
+        .insert(eserviceTemplateRiskAnalysisAnswerInReadmodelEserviceTemplate)
+        .values(riskAnalysisAnswerSQL);
     }
   });
 };
